requirements.txt
discord.py
twitchAPI